- moved DataCollectionOptions to options.py
